requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
motor>=3.1.1
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=1.10.7
jinja2>=3.1.2
python-multipart>=0.0.6
google-generativeai>=0.3.0